plan is to try: impl Remove for Option<Box<Node>>
Trying to switch to a more elegant design for the remove() function so I
don't have to duplicate code for Tree/Node due to the special case where
target == head node.  Note that for now, the new remove function is
named foo to avoid confusion while leaving the old code intact.